requirements.txt
Django>=2.2,<2.3
djangorestframework>=3.9.4,<3.10
django-environ>=0.4.5,<0.5
sentry-sdk>=0.9.0,<0.10
psycopg2>=2.8.3,<2.9 --no-binary psycopg2
requests>=2.22.0,<2.23
